* Made implicit type truncations explicit * Removed deprecated kismet math library methods * fixed some bugs caused by incorrect type conversions
#rb rob.krajcarski, ben.hoffman
#rnx

[CL 35376028 by michael galetzka in ue5-main branch]

+++ Engine/Plugins/FX/Niagara/Source/Niagara/Private/NiagaraComponent.cpp
@@ -778,10 +778,10 @@ void UNiagaraComponent::ReleaseToPool()
 uint32 UNiagaraComponent::GetApproxMemoryUsage() const
 {
 	uint32 MemoryBytes = sizeof(UNiagaraComponent);
-	MemoryBytes += OverrideParameters.GetResourceSize();
+	MemoryBytes += static_cast<uint32>(OverrideParameters.GetResourceSize());
 	if (SystemInstanceController)
 	{
-		MemoryBytes += uint32(SystemInstanceController->GetTotalBytesUsed());
+		MemoryBytes += static_cast<uint32>(SystemInstanceController->GetTotalBytesUsed());
 	}
 	return MemoryBytes;
 }
